requirements.txt
aws-cdk-lib
constructs
python-dotenv
colorlog
boto3